KTOR-2796 Refcator split ServerPlugin.kt file
