+++ ktor-server/jvm/src/io/ktor/server/application/plugins/api/KtorCallContexts.kt
@@ -29,7 +29,7 @@ public class CallContext(internal val context: PipelineContext<Unit, Application
 public class CallReceiveContext(
     private val context: PipelineContext<ApplicationReceiveRequest, ApplicationCall>
 ) : CallHandlingContext(context) {
-    public fun transformRequestBody(transform: (ByteReadChannel) -> Any) {
+    public suspend fun transformRequestBody(transform: suspend (ByteReadChannel) -> Any) {
         val receiveBody = context.subject.value as? ByteReadChannel
             ?: throw NoByteReadChannelException(context.subject.value)
         context.subject = ApplicationReceiveRequest(
@@ -46,7 +46,7 @@ public class CallReceiveContext(
 public class CallRespondContext(
     private val context: PipelineContext<Any, ApplicationCall>
 ) : CallHandlingContext(context) {
-    public fun transformResponseBody(transform: (Any) -> Any) {
+    public suspend fun transformResponseBody(transform: suspend (Any) -> Any) {
         context.subject = transform(context.subject)
     }
 }
@@ -57,7 +57,7 @@ public class CallRespondContext(
 public class CallRespondAfterTransformContext(
     private val context: PipelineContext<Any, ApplicationCall>
 ) : CallHandlingContext(context) {
-    public fun transformResponseBody(transform: (OutgoingContent) -> OutgoingContent) {
+    public suspend fun transformResponseBody(transform: suspend (OutgoingContent) -> OutgoingContent) {
         val newContent = context.subject as? OutgoingContent ?: throw NoOutgoingContentException(context.subject)
         context.subject = transform(newContent)
     }
